requirements.txt
# Core Dependencies
langgraph>=0.0.20
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.0

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Web Interface
streamlit>=1.28.0

# Utilities
typing-extensions>=4.8.0